Refactor 'now' management to a shared hook
